修改資料庫建置檔 特殊符號阻擋規則

+++ database/migrations/2021_01_13_055929_create_level_table.php
@@ -17,7 +17,7 @@ class CreateLevelTable extends Migration
             //
             $table->increments('level_id');
             $table->string('level_name', 20);
-            $table->integer('level_rank');
+            $table->tinyInteger('level_rank');
             $table->integer('level_threshold');
             $table->enum('level_status', ['Y','D']);
